requirements.txt
# Core conversion
marker-pdf>=0.3.0
python-docx>=1.1.0
pymupdf>=1.24.0
Pillow>=10.0.0

# API & UI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
gradio>=4.36.0
python-multipart>=0.0.9

# Utilities
aiofiles>=23.2.0
httpx>=0.27.0
